+++ webextension/data/js/classes/DropboxController.js
@@ -1,74 +1,126 @@
-class DropboxController {
+const DropboxController_FILENAME = 'livenotifier.json';
+
+class DropboxController extends SyncController {
 	/**
 	 * https://developer.mozilla.org/en-US/Add-ons/WebExtensions/API/identity/launchWebAuthFlow
 	 * http://dropbox.github.io/dropbox-sdk-js/tutorial-JavaScript%20SDK.html
 	 */
 
 	constructor(fileName, clientId, authToken=''){
-		this.client = null;
+		if (fileName === null || fileName === '') {
+			fileName = DropboxController_FILENAME;
+		}
+		super(fileName, clientId, authToken);
 
-		this.clientId = clientId;
-		this.authToken = authToken;
-		this.fileName = fileName;
+
+
 		this.API_ERROR_STATUS_FILE_NOT_FOUND = 409;
 		this.HTTP_STATUS_CANCEL = 499;
 	}
 
 
+
 	/**
 	 *
-	 * @param {Object} details
-	 * @return {Promise<String>}
+	 * @param {String} clientId
+	 * @param {String} authToken
+	 * @param {DropboxController} oldController
+	 * @return {?DropboxController}
 	 */
-	static identityLaunchWebAuthFlow(details){
-		return new Promise((resolve, reject) => {
-			chrome.identity.launchWebAuthFlow(details, function (responseUrl) {
-				if(typeof responseUrl!=="string" && responseUrl===undefined){
-					reject();
-				} else{
-					resolve(responseUrl);
-				}
-			})
-		})
+	static updateController(clientId, authToken, oldController=null){
+		if (clientId === '') clientId = DropboxController.d('NG93bWQyNG93bnZhZXc4');
+
+		if (oldController !== null) {
+			oldController.clientId = clientId;
+			oldController.authToken = authToken;
+		} else if (authToken !== '') {
+			oldController = new DropboxController(null, clientId, authToken);
+		} else {
+			oldController = null;
+		}
+
+		return oldController;
+	};
+
+
+
+	/*
+	 * https://developer.mozilla.org/en-US/docs/Web/API/WindowBase64/Base64_encoding_and_decoding
+	 */
+	static e(str) {
+		return btoa(encodeURIComponent(str).replace(/%([0-9A-F]{2})/g,
+			function toSolidBytes(match, p1) {
+				return String.fromCharCode('0x' + p1);
+			}));
+	}
+	static d(str) {
+		return decodeURIComponent(atob(str).split('').map(function(c) {
+			return '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2);
+		}).join(''));
+	}
+
+
+
+	/**
+	 *
+	 * @override
+	 * @return {?DropboxTypes.Dropbox}
+	 */
+	getClient(){
+		if (this.client !== null) {
+			return this.client;
+		} else if (this.clientId !== '' && this.authToken !== '') {
+			return this.client = new Dropbox.Dropbox({
+				clientId: this.clientId,
+				accessToken: DropboxController.d(this.authToken)
+			});
+		} else {
+			throw 'MissingInfo';
+		}
 	}
 
 	/**
 	 *
+	 * @override
 	 * @return {Promise<String>}
 	 */
 	async getAuthToken(){
-		if(this.authToken!==''){
+		if (this.authToken !== '') {
 			return this.authToken;
 		}
 
-		if(this.clientId===''){
+		if (this.clientId === '') {
 			return 'InvalidClientId';
 		}
 
-		const client = new Dropbox.Dropbox({"clientId": this.clientId}),
-			authUrl = client.getAuthenticationUrl(browser.identity.getRedirectURL())
-		;
+		/**
+		 *
+		 * @type {DropboxTypes.Dropbox}
+		 */
+		const client = new Dropbox.Dropbox({"clientId": this.clientId});
 
 		let urlReturned = null,
 			error
 		;
 
 		try{
+			const authUrl = client.getAuthenticationUrl(browser.identity.getRedirectURL());
+
 			urlReturned = await DropboxController.identityLaunchWebAuthFlow({url: authUrl, interactive: true});
 		} catch (e) {
 			error = e;
 		}
 
-		if(urlReturned!==null){
+		if (urlReturned !== null) {
 			const params = new URLSearchParams(new URL(urlReturned).hash.replace('#', '?'));
 
-			console.warn(urlReturned);
+			// console.warn(urlReturned);
 
-			this.authToken = params.get('access_token');
+			this.authToken = DropboxController.e(params.get('access_token'));
 
 			return this.authToken;
 		} else {
-			if(error!==undefined){
+			if (error !== undefined) {
 				consoleMsg('error', error);
 			}
 
@@ -78,23 +130,33 @@ class DropboxController {
 
 	/**
 	 *
-	 * @return {null|Dropbox.Dropbox}
+	 * @return {Promise<?boolean>}
 	 */
-	getClient(){
-		if(this.client!==null){
-			return this.client;
-		} else if(this.clientId!=='' && this.authToken!==''){
-			return this.client = new Dropbox.Dropbox({
-				clientId: this.clientId,
-				accessToken: this.authToken
-			});
-		} else {
-			throw 'MissingInfo';
+	async revokeAuthToken() {
+		const client = this.getClient();
+
+		if (client === null) {
+			return false;
 		}
+
+		let result = false;
+		try {
+			await client.authTokenRevoke();
+			result = true;
+		} catch (e) {
+			console.error(e);
+		}
+
+		return result;
 	}
+
+
+
+
 
 	/**
 	 *
+	 * @override
 	 * @return {Promise<JSON>}
 	 */
 	async get() {
@@ -111,7 +173,7 @@ class DropboxController {
 			error = e;
 		}
 
-		if(metaData!==null){
+		if (metaData !== null) {
 			try{
 				data = JSON.parse(await zDK.loadBlob(metaData.fileBlob, 'text'));
 			} catch (e) {
@@ -124,7 +186,7 @@ class DropboxController {
 				'metadata': metaData
 			};
 		} else {
-			if(error.status === this.API_ERROR_STATUS_FILE_NOT_FOUND){
+			if (error.status === this.API_ERROR_STATUS_FILE_NOT_FOUND) {
 				throw 'NoFile';
 			} else if(error!==undefined){
 				throw error;
@@ -162,6 +224,7 @@ class DropboxController {
 
 	/**
 	 *
+	 * @override
 	 * @param {JSON} jsonObject
 	 * @return {Promise<Dropbox.files.FileMetadata | boolean>}
 	 */